Machine_Learning/Desafio tripulaciones/requirements.txt
flask
transformers
pymongo
gtts
pdfminer.six
gunicorn
psycopg2-binary
PyPDF2
Flask-CORS